+++ ChangeLog.md
@@ -14,7 +14,8 @@
     - Add CParameterUser
     - Add CParameterSSH
     - Add CParameterProxy
-  - Update tigervnc to v1.13.90
+  - TigerVNC: Update tigervnc to v1.13.90
+  - RabbitVNC: Update tigervnc to v1.13.90
 
 ### v0.0.26
 - Use RabbitCommon v2